Welcome butterfly and upgraded AI!
Now we have butterfly, which is really foxy, so it is trully hard to
catch it!
And (which isn't such an amazing thing as butterfly, but anyway)
improved moving AI. Now if it is occupied, character any way try to find
the way out (it might be funny, when your character(student) bounce off
the wall :) ).
And here is new function "stay". For the case, where character tried all
the ways out and that all are occupied. It might be only if you put
character to the cage.