Winter is coming!
Now there is a north (on the right) with a snowmans and christmas trees.
And as a usual north, it is really severe. Firstly, it is separated by
mountains from others and secondly, our player (whose castle on the
south) can't stand cold. So it loses hp being on the north.
Sad, right? but not as it seems! As you know world is round. So if
you'll run from the bug having a few hp,  you can survive: there is a
fountain for you to increase hp.