Welcome your own dog! And some AI bugs fixed.
Dog! I thought Archer is really excess. And it was't interesting for me
to develop it, so I made dog :).  It is following you, but when it sees
enemy(dogs hate bugs, it is common knowledge) it attacks, but dog sees
only 3 cells around. anyway dog's helpful. Btw here is an interesting
thing, if you are dead, dog gets sad and loses hp, but then it goes to
revenge (ony by one, cold-blooded).
There was a bug, I tried hard to fix about a quarter of the day. It
appeared when i was working on dog and it's AI. So I made functions to
reduce that spagetti-like code ( now it is around, not_around). And I
noticed thad dog ignored me and went beat bugs. So it fot stuck, I
coudn't find where the problem is. So debagger saved me, it would be
better if I used it immidiatly. Now it's working nice.